refactor: move app entry to src
